for better or for worse num samples needs to be long
will need to blow away db in dev before merging

+++ src/main/java/org/cancogenvirusseq/singularity/components/hoc/InstantToArchiveBuildRequest.java
@@ -44,7 +44,7 @@ public class InstantToArchiveBuildRequest implements Function<Instant, Mono<Arch
             .status(ArchiveStatus.BUILDING)
             .type(ArchiveType.ALL)
             .hashInfo(countAndLastUpdatedResult.getLastUpdatedDate().getValueAsString())
-            .numOfSamples((int) countAndLastUpdatedResult.getNumDocuments().getValue())
+            .numOfSamples(countAndLastUpdatedResult.getNumDocuments().getValue())
             .build());
   }
